+++ src/Codeception/Command/Build.php
@@ -1,7 +1,8 @@
 <?php
 namespace Codeception\Command;
 
-use Codeception\Lib\Generator\Actor as GuyGenerator;
+use Codeception\Configuration;
+use Codeception\Lib\Generator\Actor as ActorGenerator;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -22,6 +23,11 @@ class Build extends Command
 
     protected $inheritedMethodTemplate = ' * @method void %s(%s)';
 
+    /**
+     * @var OutputInterface
+     */
+    protected $output;
+
     public function getDescription()
     {
         return 'Generates base classes for all suites';
@@ -36,20 +42,36 @@ class Build extends Command
 
 	protected function execute(InputInterface $input, OutputInterface $output)
 	{
-        $suites = $this->getSuites($input->getOption('config'));
+        $this->output = $output;
+        $this->buildActorsForConfig($input->getOption('config'));
+    }
 
-        $output->writeln("<info>Building Actor classes for suites: ".implode(', ', $suites).'</info>');
+    protected function buildActorsForConfig($configFile)
+    {
+        $config = $this->getGlobalConfig($configFile);
+        $suites = $this->getSuites($configFile);
 
+        $path = pathinfo($configFile);
+        $dir = isset($path['dirname']) ? $path['dirname'] : getcwd();
+
+        foreach ($config['include'] as $subConfig) {
+            $this->output->writeln("<comment>Included Configuration: $subConfig</comment>");
+            $this->buildActorsForConfig($dir . DIRECTORY_SEPARATOR . $subConfig);
+        }
+
+        if (!empty($suites)) {
+            $this->output->writeln("<info>Building Actor classes for suites: ".implode(', ', $suites).'</info>');
+        }
         foreach ($suites as $suite) {
-            $settings = $this->getSuiteConfig($suite, $input->getOption('config'));
-            $gen = new GuyGenerator($settings);
-            $output->writeln('<info>'.$gen->getGuy() . "</info> includes modules: ".implode(', ',$gen->getModules()));
+            $settings = $this->getSuiteConfig($suite, $configFile);
+            $gen = new ActorGenerator($settings);
+            $this->output->writeln('<info>'.Configuration::config()['namespace'].'\\'.$gen->getActorName() . "</info> includes modules: ".implode(', ',$gen->getModules()));
             $contents = $gen->produce();
 
             @mkdir($settings['path'],0755, true);
             $file = $settings['path'].$this->getClassName($settings['class_name']).'.php';
             $this->save($file, $contents, true);
-            $output->writeln("{$settings['class_name']}.php generated successfully. ".$gen->getNumMethods()." methods added");
+            $this->output->writeln("{$settings['class_name']}.php generated successfully. ".$gen->getNumMethods()." methods added");
         }
     }